refactor: [#16] AlbumView의 collectionView의 layout을 composionalLayout으로 변경
