+++ rabit/rabit/Album/AlbumViewController.swift
@@ -55,13 +55,7 @@ private extension AlbumViewController {
     }
 
     func setupAlbumCollectionView() {
-        let screenSize = UIScreen.main.bounds.size
-        let layout = UICollectionViewFlowLayout()
-        layout.itemSize = CGSize(width: screenSize.width, height: screenSize.width)
-        layout.minimumInteritemSpacing = 0
-        layout.minimumLineSpacing = 15
-        layout.headerReferenceSize = CGSize(width: screenSize.width, height: 50)
-        layout.scrollDirection = .vertical
+        let layout = getCollectionViewLayout()
 
         albumCollectionView.collectionViewLayout = layout
 
@@ -114,6 +108,12 @@ private extension AlbumViewController {
         )
 
         return dataSource
+    }
+
+    func getCollectionViewLayout() -> UICollectionViewCompositionalLayout {
+        return UICollectionViewCompositionalLayout { (section, _) -> NSCollectionLayoutSection? in
+            AlbumCollectionCompositionalLayoutFactory.shared.create()
+        }
     }
 
     func bind() {
