+++ rabit/rabit/Album/Models/Album.swift
@@ -2,18 +2,18 @@ import Foundation
 import Differentiator
 
 struct Album {
-    typealias Item = [Data]
+    typealias Item = Data
     var items: [Item]
     var date: Date
 
-    init(date: Date, items: [[Data]]) {
+    init(date: Date, items: [Item]) {
         self.date = date
         self.items = items
     }
 }
 
 extension Album: SectionModelType {
-    init(original: Album, items: [[Data]]) {
+    init(original: Album, items: [Item]) {
         self = original
         self.items = items
     }
